Modified PascalCaseWordPartsRegex to accept more "word characters".
Instead of [A-Z] this uses \p{Lu} (Upper Case Unicode letters) and instead of [a-z] this uess \p{Ll} (lower case Unicode letters). This enables us to use some word characters of different languages, see Humanizr/Humanizer#532 and Humanizr/Humanizer#563.

Note that this does not handle the example "mühldorfer STRAßE 23" from Humanizr/Humanizer#532, but if it were written as "mühldorfer STRAẞE 23", it would be okay.  The previously missing capital ß (ẞ) would require special culture dependent handling, which I do not know if it is wanted in here. However, this deals with "MÜNCHEN" and "ajouté" fine and better than the previous version.