Register Portuguese TimeSpan localisation as "pt".
According to Humanizr/Humanizer#601, the localisation is good for
Portuguese in general, not just for Brazilian Portuguese.
I copied the unit test, so that we can prove that there's no regression.

Additionally, I copied the `OrginalizeTests`, because those were
already registered as "pt", not "pt-BR".

Note that `NumberToWords` is still only valid for pt-BR unless confirmed
otherwise.