Make NumberToWords pass for Icelandic
Icelandic has some complexity that can be attributed to being 'closer'
to its root than other surviving Germanic languages and having both
more irregularities and retaining inflections such as four case
synthetic grammar. This can mostly be represented with code
branching specific to this implementation. Among other thing the
first four natural numbers change based on gender and gender must
be applied to the last number in a sequence. Rules are mostly consistent for powers of ten, but 'hundred' varies between single and
plural. After performance measurements the final version was chosen
to closely align with other implementations, while trying to be as
transparent as possible.
The Icelandic rules for ordinals are have additional requirements, e.g.
on whether the penultimate number group or only the last match the
rule.
BenchmarkDotnet used on a parallel branch to make sure the
implementation compares favarobly to other localisations.