Change th-TH culture to th
Web browsers (at least Firefox and Chromium) only send th in the
Accept-Language header, leading ASP.NET to select it as the current
culture. This leads to Humanizer failing to find the resource and
consequently falling back to the default culture. Renaming the resource
files to the more generic culture specifier fixes the issue.